Fix types and reorganise into submodules
